+++ src/applications/ivc-champva/10-10D/helpers/utilities.js
@@ -18,14 +18,6 @@ export function isInRange(val, lower, upper) {
   return val >= lower && val <= upper;
 }
 
-// Turn camelCase into capitalized words ("camelCase" => "Camel Case")
-export function makeHumanReadable(inputStr) {
-  return inputStr
-    .match(/^[a-z]+|[A-Z][a-z]*/g)
-    .map(word => word[0].toUpperCase() + word.substr(1).toLowerCase())
-    .join(' ');
-}
-
 // Helper to detect if we're on review page when we don't have access
 // to form context. Necessary because list and loop pages don't seem
 // to respect 'hideOnReview' or 'keepInPageOnReview'
