+++ src/applications/ivc-champva/shared/components/fileUploads/MissingFileOverview.jsx
@@ -6,10 +6,8 @@ import {
 import { titleUI } from 'platform/forms-system/src/js/web-component-patterns';
 import FormNavButtons from 'platform/forms-system/src/js/components/FormNavButtons';
 import PropTypes from 'prop-types';
-import {
-  identifyMissingUploads,
-  getConditionalPages,
-} from '../../helpers/supportingDocsVerification';
+import { getConditionalPages } from '../../utilities';
+import SupportingDocsVerification from './supportingDocsVerification';
 import MissingFileList from './MissingFileList';
 
 const mailInfo = (
@@ -130,6 +128,8 @@ export default function MissingFileOverview({
   showMail,
   showConsent,
   allPages,
+  fileNameMap,
+  requiredFiles,
 }) {
   const [error, setError] = useState(undefined);
   const [isChecked, setIsChecked] = useState(
@@ -137,6 +137,7 @@ export default function MissingFileOverview({
   );
   const navButtons = <FormNavButtons goBack={goBack} submitToContinue />;
   const chapters = contentAfterButtons?.props?.formConfig?.chapters;
+  const verifier = new SupportingDocsVerification(requiredFiles);
   // Create single list of pages from multiple chapter objects
   const pages =
     allPages ||
@@ -156,13 +157,13 @@ export default function MissingFileOverview({
       return checkFlags(
         conditionalPages,
         app,
-        identifyMissingUploads(conditionalPages, app, false),
+        verifier.identifyMissingUploads(conditionalPages, app, false),
       );
     });
 
   const applicantsWithMissingFiles = data.applicants
     .map(applicant => {
-      const missing = identifyMissingUploads(
+      const missing = verifier.identifyMissingUploads(
         getConditionalPages(pages, { ...data, applicants: [applicant] }, 0),
         applicant,
         false,
@@ -183,7 +184,11 @@ export default function MissingFileOverview({
     missingUploads: checkFlags(
       pages,
       data,
-      identifyMissingUploads(getConditionalPages(pages, data), data, true),
+      verifier.identifyMissingUploads(
+        getConditionalPages(pages, data),
+        data,
+        true,
+      ),
     ).missingUploads,
   };
 
@@ -250,6 +255,7 @@ export default function MissingFileOverview({
               subset="required"
               description={requiredDescription}
               disableLinks={disableLinks}
+              fileNameMap={fileNameMap}
             />
           ) : null}
           {hasReq(apps, true, showConsent) ? (
@@ -260,6 +266,7 @@ export default function MissingFileOverview({
               subset="required"
               description={requiredDescription}
               disableLinks={disableLinks}
+              fileNameMap={fileNameMap}
             />
           ) : null}
           {hasReq(sponsorMiss, false, showConsent) ? (
@@ -270,6 +277,7 @@ export default function MissingFileOverview({
               subset="optional"
               description={optionalDescription}
               disableLinks={disableLinks}
+              fileNameMap={fileNameMap}
             />
           ) : null}
           {hasReq(apps, false, showConsent) ? (
@@ -280,6 +288,7 @@ export default function MissingFileOverview({
               subset="optional"
               description={optionalDescription}
               disableLinks={disableLinks}
+              fileNameMap={fileNameMap}
             />
           ) : null}
           {requiredFilesStillMissing && showMail ? <>{mailInfo}</> : null}
@@ -315,10 +324,12 @@ MissingFileOverview.propTypes = {
   contentAfterButtons: PropTypes.object,
   data: PropTypes.object,
   disableLinks: PropTypes.bool,
+  fileNameMap: PropTypes.object,
   goBack: PropTypes.func,
   goForward: PropTypes.func,
   heading: PropTypes.node,
   optionalWarningHeading: PropTypes.node,
+  requiredFiles: PropTypes.any,
   requiredWarningHeading: PropTypes.node,
   setFormData: PropTypes.func,
   showConsent: PropTypes.bool,
